% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMARETTOinit_all}
\alias{AMARETTOinit_all}
\title{AMARETTOinit_all}
\format{List}
\source{
\url{CommunityAMARETTO}
}
\usage{
AMARETTOinit_all
}
\description{
A dataframe containing of MSIGDB geneset names, their urls and descriptions.
}
\keyword{datasets}
