% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMARETTOresults_all}
\alias{AMARETTOresults_all}
\title{AMARETTOresults_all}
\format{An object of class \code{list} of length 2.}
\source{
\url{CommunityAMARETTO}
}
\usage{
AMARETTOresults_all
}
\description{
A dataframe containing of MSIGDB geneset names, their urls and descriptions.
}
\keyword{datasets}
