% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_ColorOneModule.R
\name{cAMARETTO_ColorOneModule}
\alias{cAMARETTO_ColorOneModule}
\title{cAMARETTO_ColorOneModule}
\usage{
cAMARETTO_ColorOneModule(cAMARETTOnetworkM, cAMARETTOnetworkC, ModuleNr)
}
\arguments{
\item{ModuleNr}{}
}
\value{
a plot with one (or none) colored modules
}
\description{
cAMARETTO_ColorOneModule
}
