% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{cAMARETTO_Cytoscape}
\alias{cAMARETTO_Cytoscape}
\title{Title cAMARETTO_Cytoscape}
\usage{
cAMARETTO_Cytoscape(cAMARETTOsList, communityReportURL = "",
  cytoscape_name = "my_cytoscape")
}
\arguments{
\item{cAMARETTOsList}{A list containing Community AMARETTO results with the following format : list(cAMARETTOresults = cAMARETTOresults,cAMARETTOnetworkM = cAMARETTOnetworkM, cAMARETTOnetworkC = cAMARETTOnetworkC)}

\item{communityReportURL}{Optional URL linked to the Community-AMARETTO HTML report index page.}

\item{cytoscape_name}{A character, for naming the network which will be shown in cytoscape.}
}
\value{
result
}
\description{
Title cAMARETTO_Cytoscape
}
\examples{
 
try(
cytoscape_name<-"cAMARETTO_Liver2DS"
cAMARETTOsList<-readRDS(file="./outputs/cAMARETTO_Liver2DS.rds")
communityReportURL<-"http://portals.broadinstitute.org/pochetlab/demo/cAMARETTO_Liver_2DS/" 
cAMARETTO_Cytoscape(cAMARETTOsList,communityReportURL = "",cytoscape_name="my_cytoscape")
)
}
