% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_DataReadDownload.R
\name{cAMARETTO_HTML_Read}
\alias{cAMARETTO_HTML_Read}
\title{Title cAMARETTO_HTML_Read}
\usage{
cAMARETTO_HTML_Read(HTMLsAMARETTOlist, unzipParentDirectory = getwd())
}
\arguments{
\item{HTMLsAMARETTOlist}{a list of directories for each AMARETTO HTML report.}

\item{unzipParentDirectory}{a directory address where the html report files are unzipped to}
}
\value{
a named vector with directories of unzipped HTML reprorts.
}
\description{
Title cAMARETTO_HTML_Read
}
\examples{
cAMARETTO_HTML_Read(list(TCGA_LIHC="TCGA_LIHC_report.zip",TCGA_GBM="TCGA_GBM_report.zip"))
}
