% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_identifycom.R
\name{cAMARETTO_IdentifyCom}
\alias{cAMARETTO_IdentifyCom}
\title{cAMARETTO_IdentifyCom}
\usage{
cAMARETTO_IdentifyCom(cAMARETTOnetworkM, color_list = NULL,
  filterComm = TRUE, ratioCommSize = 0.01, MinRuns = 2,
  ratioRunSize = 0.1, ratioEdgesInOut = 0.5)
}
\arguments{
\item{color_list}{An optional list with colors}

\item{edge_method}{Define edge weights based p-values or overlap between the gene sets.}
}
\value{
a list with the module network, layout for the network, used p-value, used overlap en colors
}
\description{
cAMARETTO_IdentifyCom
}
