% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_InformationTable.R
\name{cAMARETTO_InformationTable}
\alias{cAMARETTO_InformationTable}
\title{cAMARETTO_InformationTable}
\usage{
cAMARETTO_InformationTable(cAMARETTOnetworkM, cAMARETTOnetworkC)
}
\arguments{
\item{cAMARETTOnetworkM}{The output of the Module Network function.}

\item{cAMARETTOnetworkC}{The output of the Identify Communities function.}
}
\value{
an information table about the communities containing the included nodes and overlapping genes per community.
}
\description{
cAMARETTO_InformationTable
}
\examples{

cAMARETTO_InformationTable(cAMARETTOnetworkM, cAMARETTOnetworkC)

}
