% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_modulenetwork.R
\name{cAMARETTO_ModuleNetwork}
\alias{cAMARETTO_ModuleNetwork}
\title{cAMARETTO_ModuleNetwork}
\usage{
cAMARETTO_ModuleNetwork(cAMARETTOresults, pvalue, inter,
  color_list = NULL, edge_method = "pvalue")
}
\arguments{
\item{cAMARETTOresults}{}

\item{pvalue}{pvalue cut-off for each gene set}

\item{inter}{minimal overlap between two gene sets}

\item{color_list}{An optional list with colors}

\item{edge_method}{Define edge weights based p-values or overlap between the gene sets.}
}
\value{
a list with the module network, layout for the network, used p-value, used overlap en colors
}
\description{
cAMARETTO_ModuleNetwork
}
