% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_DataReadDownload.R
\name{cAMARETTO_Read}
\alias{cAMARETTO_Read}
\title{cAMARETTO_Read}
\usage{
cAMARETTO_Read(AMARETTOdirectories)
}
\arguments{
\item{AMARETTOdirectories}{}
}
\value{
a list with the AMARETTOinit and AMARETTOresults
}
\description{
cAMARETTO_Read
}
