% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_ImportRuns.R
\name{cAMARETTO_ReadZipExport}
\alias{cAMARETTO_ReadZipExport}
\title{cAMARETTO_ReadZipExport}
\usage{
cAMARETTO_ReadZipExport(zipdirs)
}
\arguments{
\item{zipdirs}{A list of multiple AMARETTO results in a zipformat retrieved by the AMARETTO_ExportData function}
}
\value{
a list with AMARETTOinit and AMARETTOresults data objects from multiple runs
}
\description{
cAMARETTO_ReadZipExport
}
