% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCLE_Liver_data}
\alias{CCLE_Liver_data}
\title{CCLE_Liver_data}
\format{List}
\source{
\url{CommunityAMARETTO}
}
\usage{
CCLE_Liver_data
}
\description{
A list containing data from an analysis for CCLE-Liver.
The list include AMARETTO_result object calculated from AMARETTO
analysis.
}
\keyword{datasets}
