% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{CommunityHyperLink}
\alias{CommunityHyperLink}
\title{CommunityHyperLink}
\usage{
CommunityHyperLink(Community, Community_key, Community_type)
}
\arguments{
\item{Community}{CommunityHyperLink}

\item{Community_key}{Community_key}

\item{Community_type}{Community_type}
}
\value{
a hyperlink and presentable name for
the communities used for datatables
}
\description{
CommunityHyperLink
}
\examples{
try(CommunityHyperLink(Community,Community_key,Community_type))
}
