% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{CommunityModuleTableCreate}
\alias{CommunityModuleTableCreate}
\title{Title CommunityModuleTableCreate}
\usage{
CommunityModuleTableCreate(cAMARETTOresults, cAMARETTOnetworkM,
  cAMARETTOnetworkC, RunInfo)
}
\arguments{
\item{cAMARETTOresults}{}

\item{cAMARETTOnetworkM}{}

\item{cAMARETTOnetworkC}{}

\item{RunInfo}{}
}
\value{
Community To Module Dataframe
}
\description{
Title CommunityModuleTableCreate
}
\examples{
CommunityModuleTableCreate (cAMARETTOresults, cAMARETTOnetworkM, cAMARETTOnetworkC, RunInfo)

}
