% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{ComRunModGenInfo}
\alias{ComRunModGenInfo}
\title{Title ComRunModGenInfo}
\usage{
ComRunModGenInfo(cAMARETTOresults, cAMARETTOnetworkM, cAMARETTOnetworkC)
}
\arguments{
\item{cAMARETTOresults}{}

\item{cAMARETTOnetworkM}{}

\item{cAMARETTOnetworkC}{}
}
\value{
a dataframe contaning all communities, runname, and modules relationships.
}
\description{
Title ComRunModGenInfo
}
\examples{
ComRunModGenInfo(cAMARETTOresults,cAMARETTOnetworkM,cAMARETTOnetworkC)
}
