% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{CreatePhenotypeTable}
\alias{CreatePhenotypeTable}
\title{Title CreatePhenotypeTable}
\usage{
CreatePhenotypeTable(cAMARETTOresults, cAMARETTOnetworkM,
  cAMARETTOnetworkC, PhenotypeTablesList)
}
\arguments{
\item{cAMARETTOresults}{}

\item{cAMARETTOnetworkM}{}

\item{cAMARETTOnetworkC}{}

\item{PhenotypeTablesList}{}
}
\value{
results
}
\description{
Title CreatePhenotypeTable
}
\examples{
try(CreatePhenotypeTable(cAMARETTOresults, cAMARETTOnetworkM, cAMARETTOnetworkC, PhenotypeTables))
}
