% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Driver_Genes}
\alias{Driver_Genes}
\title{Driver_Genes}
\format{An object of class \code{list} of length 10.}
\source{
\url{CommunityAMARETTO}
}
\usage{
Driver_Genes
}
\description{
A dataframe containing of MSIGDB geneset names, their urls and descriptions.
}
\keyword{datasets}
