% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{drivers_communities_summary}
\alias{drivers_communities_summary}
\title{Title drivers_communities_summary}
\usage{
drivers_communities_summary(
  cAMARETTOresults,
  cAMARETTOnetworkM,
  cAMARETTOnetworkC
)
}
\arguments{
\item{cAMARETTOresults}{cAMARETTOresults object}

\item{cAMARETTOnetworkM}{cAMARETTOnetworkM object}

\item{cAMARETTOnetworkC}{cAMARETTOnetworkC object}
}
\value{
a list containing 3 tables summarising drivers-comunities relationships.
}
\description{
Title drivers_communities_summary
}
\examples{
try(drivers_communities_summary(cAMARETTOresults,cAMARETTOnetworkM,cAMARETTOnetworkC))
}
