% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{DriversSharedTbl}
\alias{DriversSharedTbl}
\title{Title DriversSharedTbl}
\usage{
DriversSharedTbl(cAMARETTOresults, cAMARETTOnetworkM, cAMARETTOnetworkC)
}
\arguments{
\item{cAMARETTOresults}{}

\item{cAMARETTOnetworkM}{}

\item{cAMARETTOnetworkC}{}
}
\value{
Frequency of driver genes across different communities and datasets
}
\description{
Title DriversSharedTbl
}
\examples{
 DriversSharedTbl(cAMARETTOresults, cAMARETTOnetworkM, cAMARETTOnetworkC)
}
