% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_Results.R
\name{Extract_Genes_Modules_All}
\alias{Extract_Genes_Modules_All}
\title{Title Extract_Genes_Modules_All}
\usage{
Extract_Genes_Modules_All(AMARETTOresults_all, gmt_filelist)
}
\arguments{
\item{AMARETTOresults_all}{}

\item{gmt_filelist}{}
}
\value{
a dataframe for all AMARETTO files and given gmt file with the following structure :  runname, ModuleNr, genename, gene-types, and weights of the driver genes.
}
\description{
Title Extract_Genes_Modules_All
}
\examples{
try(Extract_Genes_Modules_All(AMARETTOresults_all,gmt_filelist))
}
