% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_Results.R
\name{GmtFromModules}
\alias{GmtFromModules}
\title{GmtFromModules}
\usage{
GmtFromModules(AMARETTOinit, AMARETTOresults, gmt_file, run)
}
\arguments{
\item{gmt_file}{A gmtfilename}

\item{run}{A runname}

\item{AMARETTOinit_all}{A list of multiple AMARETTO_Initialize outputs. The names are run names.}

\item{AMARETTOresults_all}{A list of multiple AMARETTO_Run outputs. The names are run names.}
}
\value{
Creates a gmt file for a AMARETTO run
}
\description{
GmtFromModules
}
