% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{HGTGeneEnrichmentList}
\alias{HGTGeneEnrichmentList}
\title{HGTGeneEnrichmentList

Calculates the p-values for unranked gene set enrichment based on two gmt files as input and the hyper geometric test.}
\usage{
HGTGeneEnrichmentList(genelist, gmtfile, NrCores, ref.numb.genes = 45956)
}
\arguments{
\item{genelist}{The gmt file with reference gene set.}

\item{gmtfile}{The gmt file with gene sets to test. In our case, the gmt file of the modules.}

\item{NrCores}{Number of cores used for parallelization.}

\item{ref.numb.genes}{The total number of genes teste, standard equal to 45956 (MSIGDB standard).}
}
\value{
a list a dataframe for Hypergeometric test
}
\description{
HGTGeneEnrichmentList

Calculates the p-values for unranked gene set enrichment based on two gmt files as input and the hyper geometric test.
}
\examples{
try(HGTGeneEnrichmentList(genelist, gmtfile, NrCores, ref.numb.genes = 45956))

}
\keyword{internal}
