% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{HubAuthority_scores_CommunityAMARETTO}
\alias{HubAuthority_scores_CommunityAMARETTO}
\title{Title HubAuthority_scores_CommunityAMARETTO}
\usage{
HubAuthority_scores_CommunityAMARETTO(
  cAMARETTOresults,
  cAMARETTOnetworkM,
  cAMARETTOnetworkC
)
}
\arguments{
\item{cAMARETTOresults}{cAMARETTOresults object}

\item{cAMARETTOnetworkM}{cAMARETTOnetworkM object}

\item{cAMARETTOnetworkC}{cAMARETTOnetworkC object}
}
\value{
a list containing Hub-score authority-score of genes and modules.
}
\description{
Title HubAuthority_scores_CommunityAMARETTO
}
\examples{
try(Hubscore_CommunityAMARETTO(cAMARETTOresults,cAMARETTOnetworkM,cAMARETTOnetworkC))
}
