% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_results.R
\name{HyperGTestGeneEnrichment}
\alias{HyperGTestGeneEnrichment}
\title{GmtFromModules}
\usage{
HyperGTestGeneEnrichment(gmtfile, testgmtfile, NrCores,
  ref.numb.genes = 45956)
}
\arguments{
\item{ref.numb.genes}{The reference number of genes.}

\item{gmtfile1}{A gmtfilename that you want to compare}

\item{gmtfile2}{A second gmtfile to compare with.}
}
\value{
Creates resultfile with p-values and padj when comparing two gmt files with a hyper geometric test.
}
\description{
GmtFromModules
}
