% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{InitialCheckInputs}
\alias{InitialCheckInputs}
\title{Title InitialCheckInputs}
\usage{
InitialCheckInputs(cAMARETTOresults, output_address, HTMLsAMARETTOlist,
  CopyAMARETTOReport, hyper_geo_reference, hyper_geo_reference_gp,
  hyper_geo_reference_cp)
}
\arguments{
\item{cAMARETTOresults}{}

\item{output_address}{}

\item{HTMLsAMARETTOlist}{}

\item{CopyAMARETTOReport}{}

\item{hyper_geo_reference}{}

\item{hyper_geo_reference_gp}{}

\item{hyper_geo_reference_cp}{}
}
\value{
RunInfo dataframe
}
\description{
Title InitialCheckInputs
}
\examples{
try(
InitialCheckInputs(cAMARETTOresults,output_address="./",HTMLsAMARETTOlist,CopyAMARETTOReport=FALSE,hyper_geo_reference)
)
}
