% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MA_matrix_CCLE}
\alias{MA_matrix_CCLE}
\title{MA_matrix_CCLE}
\format{An object of class \code{matrix} with 13840 rows and 25 columns.}
\source{
\url{CommunityAMARETTO}
}
\usage{
MA_matrix_CCLE
}
\description{
A dataframe containing of MSIGDB geneset names, their urls and descriptions.
}
\keyword{datasets}
