% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MA_matrix_LIHC}
\alias{MA_matrix_LIHC}
\title{MA_matrix_LIHC}
\format{An object of class \code{matrix} with 14906 rows and 40 columns.}
\source{
\url{CommunityAMARETTO}
}
\usage{
MA_matrix_LIHC
}
\description{
A dataframe containing of MSIGDB geneset names, their urls and descriptions.
}
\keyword{datasets}
