% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MET_matrix_LIHC}
\alias{MET_matrix_LIHC}
\title{MET_matrix_LIHC}
\format{An object of class \code{matrix} with 2780 rows and 40 columns.}
\source{
\url{CommunityAMARETTO}
}
\usage{
MET_matrix_LIHC
}
\description{
A dataframe containing of MSIGDB geneset names, their urls and descriptions.
}
\keyword{datasets}
