% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_Results.R
\name{readGMT}
\alias{readGMT}
\title{readGMT}
\usage{
readGMT(filename)
}
\arguments{
\item{filename}{A gmtfilename}
}
\value{
Reads a gmt file
}
\description{
readGMT
}
\examples{
readGMT("file.gmt")
}
