% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{RunHyperLink}
\alias{RunHyperLink}
\title{Title RunHyperLink}
\usage{
RunHyperLink(
  Run_Names,
  AMARETTOres,
  HTMLsAMARETTOlist,
  CopyAMARETTOReport,
  page = 1
)
}
\arguments{
\item{Run_Names}{a vector of Run names}

\item{AMARETTOres}{list of AMARETTOres object}

\item{HTMLsAMARETTOlist}{A list with AMARETTO reports to link with the Community AMARETTO report. If NULL, no links are added.}

\item{CopyAMARETTOReport}{Boolean to indicate if the AMARETTO reports needs to be copied in the AMARETTO report directory. In this way links are contained when moving the HTML directory.}

\item{page}{1 for index page and 2 for community page}
}
\value{
Hyperlinks for Run Names
}
\description{
Title RunHyperLink
}
\examples{
try(RunHyperLink(Run_Names, AMARETTOres, HTMLsAMARETTOlist, CopyAMARETTOReport, page=1))

}
