% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scHBV_data}
\alias{scHBV_data}
\title{scHBV_data}
\format{List}
\source{
\url{CommunityAMARETTO}
}
\usage{
scHBV_data
}
\description{
A list containing data from an analysis for single cell HBV.
The list include AMARETTO_result object calculated from AMARETTO
analysis.
}
\keyword{datasets}
