% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scHCV_data}
\alias{scHCV_data}
\title{scHCV_data}
\format{List}
\source{
\url{CommunityAMARETTO}
}
\usage{
scHCV_data
}
\description{
A list containing data from an analysis for single cell HCV.
The list include AMARETTO_result object calculated from AMARETTO
analysis.
}
\keyword{datasets}
