% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCGA_LIHC_data}
\alias{TCGA_LIHC_data}
\title{TCGA_LIHC_data}
\format{List}
\source{
\url{CommunityAMARETTO}
}
\usage{
TCGA_LIHC_data
}
\description{
A list containing data from an analysis for TCGA-LIHC.
The list include following objects: 1) AMARETTO_result calculated
from AMARETTO analysis. 2) Phenotype Association analysis table.
}
\keyword{datasets}
