% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_heatmap.R
\name{cAMARETTO_heatmap}
\alias{cAMARETTO_heatmap}
\title{cAMARETTO_heatmap}
\usage{
cAMARETTO_heatmap(cAMARETTOresults, run1, run2)
}
\arguments{
\item{cAMARETTOresults}{cAMARETTO_Run output}

\item{run1}{name of run 1}

\item{run2}{name of run 2}
}
\value{
a heatmap with the comparison between two runs
}
\description{
cAMARETTO_heatmap
}
