% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{cAMARETTO_HTMLreport}
\alias{cAMARETTO_HTMLreport}
\title{cAMARETTO_HTMLreport}
\usage{
cAMARETTO_HTMLreport(cAMARETTOresults, cAMARETTOnetworkM,
  cAMARETTOnetworkC, report_address = "./")
}
\arguments{
\item{report_address}{}
}
\value{
a plot with one (or none) colored modules
}
\description{
cAMARETTO_HTMLreport
}
