% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{cAMARETTO_HTMLreport}
\alias{cAMARETTO_HTMLreport}
\title{cAMARETTO_HTMLreport}
\usage{
cAMARETTO_HTMLreport(cAMARETTOresults, cAMARETTOnetworkM,
  cAMARETTOnetworkC, output_address = "./", HTMLsAMARETTOlist = NULL,
  CopyAMARETTOReport = TRUE, hyper_geo_test_bool = FALSE,
  hyper_geo_reference = NULL, MSIGDB = FALSE, GMTURL = FALSE,
  NrCores = 2)
}
\arguments{
\item{NrCores}{}
}
\value{
A set of HTMLs, giving caracteristics of the communities
}
\description{
cAMARETTO_HTMLreport
}
