% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_InformationTable.R
\name{cAMARETTO_InformationTable}
\alias{cAMARETTO_InformationTable}
\title{cAMARETTO_InformationTable}
\usage{
cAMARETTO_InformationTable(cAMARETTOnetworkM, cAMARETTOnetworkC)
}
\arguments{
\item{ModuleNr}{}
}
\value{
a plot with one (or none) colored modules
}
\description{
cAMARETTO_InformationTable
}
