% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_modulenetwork.R
\name{cAMARETTO_ModuleNetwork}
\alias{cAMARETTO_ModuleNetwork}
\title{cAMARETTO_ModuleNetwork

Creates a module network.}
\usage{
cAMARETTO_ModuleNetwork(cAMARETTOresults, pvalue = 0.05, inter = 5,
  color_list = NULL, edge_method = "pvalue", plot_network = TRUE)
}
\arguments{
\item{cAMARETTOresults}{}

\item{pvalue}{pvalue cut-off for each gene set}

\item{inter}{minimal overlap between two gene sets}

\item{color_list}{An optional list with colors}

\item{edge_method}{Define edge weights based p-values or overlap between the gene sets.}

\item{plot_network}{If TRUE, plots the Community Network at the end.}
}
\value{
a list with the module network, layout for the network, used p-value, used overlap en colors
}
\description{
cAMARETTO_ModuleNetwork

Creates a module network.
}
\examples{
try(
cAMARETTOnetworkM<-cAMARETTO_ModuleNetwork(cAMARETTOresults,0.10,5)
)
}
