% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_results.R
\name{cAMARETTO_Results}
\alias{cAMARETTO_Results}
\title{cAMARETTO_Results}
\usage{
cAMARETTO_Results(AMARETTOinit_all, AMARETTOresults_all, nrCores = 1,
  output_dir = "./", gmt_filelist = NULL)
}
\arguments{
\item{AMARETTOinit_all}{A list of multiple AMARETTO_Initialize outputs. The names are run names.}

\item{AMARETTOresults_all}{A list of multiple AMARETTO_Run outputs. The names are run names.}

\item{output_dir}{A directory that stores gmt files for the runs}

\item{parallelparam}{BiocParallel BPPARAM}
}
\value{
a list with AMARETTOinit and AMARETTOresults data objects from multiple runs
}
\description{
cAMARETTO_Results
}
