% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{CommunityModuleTableCreate}
\alias{CommunityModuleTableCreate}
\title{CommunityModuleTableCreate}
\usage{
CommunityModuleTableCreate(
  cAMARETTOresults,
  cAMARETTOnetworkM,
  cAMARETTOnetworkC,
  HTMLsAMARETTOlist,
  CopyAMARETTOReport
)
}
\arguments{
\item{cAMARETTOresults}{cAMARETTOresults object}

\item{cAMARETTOnetworkM}{cAMARETTOnetworkM object}

\item{cAMARETTOnetworkC}{cAMARETTOnetworkC object}

\item{HTMLsAMARETTOlist}{List contating the address to index page for each data}

\item{CopyAMARETTOReport}{Boolean to indicate if the AMARETTO reports needs to be copied in the AMARETTO report directory. In this way links are contained when moving the HTML directory.}
}
\value{
Community To Module Dataframe
}
\description{
CommunityModuleTableCreate
}
\examples{
try( 
CommunityModuleTableCreate (cAMARETTOresults, cAMARETTOnetworkM, cAMARETTOnetworkC, RunInfo)
)
}
