% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{ComRunModGenInfo}
\alias{ComRunModGenInfo}
\title{Title ComRunModGenInfo}
\usage{
ComRunModGenInfo(cAMARETTOresults, cAMARETTOnetworkM, cAMARETTOnetworkC)
}
\arguments{
\item{cAMARETTOresults}{cAMARETTOresults object}

\item{cAMARETTOnetworkM}{cAMARETTOnetworkM object}

\item{cAMARETTOnetworkC}{cAMARETTOnetworkC object}
}
\value{
a dataframe contaning all communities, runname, and modules relationships.
}
\description{
Title ComRunModGenInfo
}
\examples{

try(
df<-ComRunModGenInfo(cAMARETTOresults,cAMARETTOnetworkM,cAMARETTOnetworkC)
)

}
