% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{CreateHyperGeoTestAll}
\alias{CreateHyperGeoTestAll}
\title{Title CreateHyperGeoTestAll}
\usage{
CreateHyperGeoTestAll(cAMARETTOresults, cAMARETTOnetworkM,
  cAMARETTOnetworkC, hyper_geo_reference, driverGSEA = TRUE,
  NrCores = 4)
}
\arguments{
\item{cAMARETTOresults}{}

\item{cAMARETTOnetworkM}{}

\item{cAMARETTOnetworkC}{}

\item{hyper_geo_reference}{}

\item{driverGSEA}{}

\item{NrCores}{}
}
\value{
Geneset Enrichment Analysis for the entire communities.
}
\description{
Title CreateHyperGeoTestAll
}
\examples{
try(
CreateHyperGeoTestAll(cAMARETTOresults, cAMARETTOnetworkM, cAMARETTOnetworkC, hyper_geo_reference = './h.all.v6.2.symbols.gmt', driverGSEA=TRUE,NrCores=4)
)
}
