% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{CreatePhenotypeTable}
\alias{CreatePhenotypeTable}
\title{CreatePhenotypeTable}
\usage{
CreatePhenotypeTable(
  cAMARETTOresults,
  cAMARETTOnetworkM,
  cAMARETTOnetworkC,
  PhenotypeTablesList
)
}
\arguments{
\item{cAMARETTOresults}{list of cAMARETTOresults results}

\item{cAMARETTOnetworkM}{cAMARETTOnetworkM object}

\item{cAMARETTOnetworkC}{cAMARETTOnetworkC object}

\item{PhenotypeTablesList}{a list of phenotypeassociation tables
computed in AMARETTO aalysis}
}
\value{
results
}
\description{
CreatePhenotypeTable
}
\examples{
try(CreatePhenotypeTable(cAMARETTOresults,
   cAMARETTOnetworkM,
   cAMARETTOnetworkC,
   PhenotypeTables))
}
