% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_Results.R
\name{GmtFromModules}
\alias{GmtFromModules}
\title{GmtFromModules}
\usage{
GmtFromModules(AMARETTOresults, gmt_file, run, Drivers = FALSE)
}
\arguments{
\item{AMARETTOresults}{A AMARETTO_Run output.}

\item{gmt_file}{A gmtfilename}

\item{run}{A runname}

\item{Drivers}{Add drivers to the gmt-file}
}
\value{
Creates a gmt file for a AMARETTO run
}
\description{
GmtFromModules
}
