% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{InitialCheckInputs}
\alias{InitialCheckInputs}
\title{Title InitialCheckInputs}
\usage{
InitialCheckInputs(cAMARETTOresults, output_address, HTMLsAMARETTOlist,
  CopyAMARETTOReport, hyper_geo_test_bool, hyper_geo_reference)
}
\arguments{
\item{cAMARETTOresults}{}

\item{output_address}{}

\item{HTMLsAMARETTOlist}{}

\item{CopyAMARETTOReport}{}

\item{hyper_geo_test_bool}{}

\item{hyper_geo_reference}{}
}
\value{
RunInfo dataframe
}
\description{
Title InitialCheckInputs
}
\examples{
InitialCheckInputs(cAMARETTOresults,output_address="./",HTMLsAMARETTOlist,CopyAMARETTOReport=FALSE,hyper_geo_test_bool = TRUE,hyper_geo_reference)
}
