% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{ModuleHyperLink}
\alias{ModuleHyperLink}
\title{Title ModuleHyperLink}
\usage{
ModuleHyperLink(
  Module,
  Run_Names,
  AMARETTOres,
  HTMLsAMARETTOlist,
  CopyAMARETTOReport,
  page = 1
)
}
\arguments{
\item{Module}{Module Nr ex. 4}

\item{Run_Names}{Run_Names ex. "TCGA_LIHC"}

\item{AMARETTOres}{}

\item{HTMLsAMARETTOlist}{}

\item{CopyAMARETTOReport}{Boolean to indicate if the AMARETTO reports needs to be copied in the AMARETTO report directory. In this way links are contained when moving the HTML directory.}

\item{page}{1 for index pages and 2 for community pages}
}
\value{
Hyperlinks for Modules
}
\description{
Title ModuleHyperLink
}
\examples{
try(ModuleHyperLink(Module,Run_Names,AMARETTOres,HTMLsAMARETTOlist,CopyAMARETTOReport,page=1))
}
