% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{ModuleHyperLink}
\alias{ModuleHyperLink}
\title{Title ModuleHyperLink}
\usage{
ModuleHyperLink(Module, Run_Names, AMARETTOres, HTMLsAMARETTOlist,
  CopyAMARETTOReport, page = 1)
}
\arguments{
\item{Module}{}

\item{Run_Names}{}

\item{AMARETTOres}{}

\item{HTMLsAMARETTOlist}{}

\item{CopyAMARETTOReport}{}

\item{page}{}
}
\value{
Hyperlinks for Modules
}
\description{
Title ModuleHyperLink
}
\examples{
ModuleHyperLink(Module,Run_Names,AMARETTOres,HTMLsAMARETTOlist,CopyAMARETTOReport,page=1)
}
