% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MsigdbMapping}
\alias{MsigdbMapping}
\title{MsigdbMapping}
\format{List}
\source{
\url{CommunityAMARETTO}
}
\usage{
MsigdbMapping
}
\description{
A dataframe containing of MSIGDB geneset names, their urls and descriptions.
}
\keyword{datasets}
