% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cAMARETTO_HTMLreport.R
\name{RunHyperLink}
\alias{RunHyperLink}
\title{Title RunHyperLink}
\usage{
RunHyperLink(Run_Names, AMARETTOres, HTMLsAMARETTOlist, CopyAMARETTOReport,
  page = 1)
}
\arguments{
\item{Run_Names}{}

\item{AMARETTOres}{}

\item{HTMLsAMARETTOlist}{}

\item{CopyAMARETTOReport}{}

\item{page}{1 for index page and 2 for community page}
}
\value{
Hyperlinks for Run Names
}
\description{
Title RunHyperLink
}
\examples{
RunHyperLink(Run_Names, AMARETTOres, HTMLsAMARETTOlist, CopyAMARETTOReport, page=1)

}
